backend/requirements.txt
Flask==3.0.3
flask-cors==4.0.1
openai==0.28.0
python-dotenv==1.0.0
